requirements.txt
django==3.1.14
djongo==1.3.6
sqlparse==0.2.4
pymongo>=3.12.3
dnspython==2.3.0
asgiref==3.8.1
pytz

djangorestframework==3.14.0
django-cors-headers==3.7.0  # ✅ phù hợp Django 3.1.x

# ✅ Cập nhật để phù hợp crispy-bootstrap5
django-crispy-forms==1.13.0
crispy-bootstrap5==0.5
django-js-asset==2.0

feedparser>=6.0.10
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0
beautifulsoup4>=4.12.2
nltk>=3.8.1
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
kafka-python>=2.0.2
